% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data_.R
\docType{data}
\name{hist_navs}
\alias{hist_navs}
\title{Historic monthly index values for multiple USD asset classes}
\format{A data frame with 297 rows and 25 variables}
\source{
Bloomberg
}
\usage{
hist_navs
}
\description{
A dataset containing monthly USD multi-asset index values from
December 1993 - August 2018
}
\keyword{datasets}
