% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data_.R
\docType{data}
\name{hist_rets}
\alias{hist_rets}
\title{Historic monthly returns for multiple USD asset classes}
\format{A data frame with 296 rows and 25 variables}
\source{
Bloomberg, lumiica
}
\usage{
hist_rets
}
\description{
A dataset containing monthly USD multi-asset returns from
December 1993 - August 2018
}
\keyword{datasets}
