% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data_.R
\docType{data}
\name{hist_sims}
\alias{hist_sims}
\title{Simulated annual returns for multiple USD asset classes}
\format{A data frame with 1000 rows and 25 variables}
\source{
Bloomberg, lumiica
}
\usage{
hist_sims
}
\description{
A dataset containing 1000 simulated annual USD multi-asset returns
based on monthly data from December 1993 - August 2018
}
\keyword{datasets}
