% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data_.R
\docType{data}
\name{histrets}
\alias{histrets}
\title{Historic monthly index values for multiple USD asset classes}
\format{A data frame with 297 rows and 25 variables}
\source{
Bloomberg
}
\usage{
histrets
}
\description{
A dataset containing monthly USD multi-asset index values from
December 1993 - August 2018
* SAME AS us_multi_hist, for legacy code purposes
}
\keyword{datasets}
