% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_multi_hist}
\alias{us_multi_hist}
\title{Historic monthly index values for multiple USD asset classes}
\format{A data frame with 297 rows and 25 variables}
\source{
Bloomberg
}
\usage{
us_multi_hist
}
\description{
A dataset containing monthly USD multi-asset index values from
December 1993 - August 2018
}
\keyword{datasets}
